feat: refactor location handling by modularizing queries and improving slug generation logic
