+++ lib/db/schema/location.ts
@@ -1,3 +1,5 @@
+import type { z } from 'zod';
+
 import { int, real, sqliteTable, text, unique } from 'drizzle-orm/sqlite-core';
 import { createInsertSchema } from 'drizzle-zod';
 
@@ -29,3 +31,5 @@ export const InsertLocation = createInsertSchema(location, {
   createdAt: true,
   updatedAt: true,
 });
+
+export type InsertLocationType = z.infer<typeof InsertLocation>;
